# backend/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"MessageResponse": {
				"description": "Generic success envelope",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LogbookEntryBody": {
				"properties": {
					"frequency": {
						"type": "string",
						"description": "DAILY | WEEKLY"
					},
					"entryDate": {
						"type": "string",
						"description": "ISO 8601 date string"
					},
					"weekNumber": {
						"type": "number",
						"format": "double"
					},
					"activitiesDone": {
						"type": "string"
					},
					"skillsGained": {
						"type": "string"
					},
					"challenges": {
						"type": "string"
					},
					"nextWeekPlan": {
						"type": "string"
					},
					"enrollmentId": {
						"type": "string"
					}
				},
				"required": [
					"frequency",
					"entryDate",
					"activitiesDone"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_LogbookEntryBody_": {
				"properties": {
					"frequency": {
						"type": "string",
						"description": "DAILY | WEEKLY"
					},
					"entryDate": {
						"type": "string",
						"description": "ISO 8601 date string"
					},
					"weekNumber": {
						"type": "number",
						"format": "double"
					},
					"activitiesDone": {
						"type": "string"
					},
					"skillsGained": {
						"type": "string"
					},
					"challenges": {
						"type": "string"
					},
					"nextWeekPlan": {
						"type": "string"
					},
					"enrollmentId": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ApproveEntryBody": {
				"properties": {
					"note": {
						"type": "string",
						"description": "Optional supervisor note"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AttendanceBody": {
				"properties": {
					"date": {
						"type": "string",
						"description": "ISO 8601 date string"
					},
					"status": {
						"type": "string",
						"description": "PRESENT | ABSENT | LATE | EXCUSED"
					},
					"checkInTime": {
						"type": "string"
					},
					"checkOutTime": {
						"type": "string"
					},
					"hoursLogged": {
						"type": "number",
						"format": "double"
					},
					"note": {
						"type": "string"
					},
					"enrollmentId": {
						"type": "string"
					}
				},
				"required": [
					"date",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InternshipBody": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"description": "ACADEMIC | PROFESSIONAL"
					},
					"field": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"isRemote": {
						"type": "boolean"
					},
					"isPaid": {
						"type": "boolean"
					},
					"stipendAmount": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string"
					},
					"durationWeeks": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"applicationDeadline": {
						"type": "string"
					},
					"slots": {
						"type": "number",
						"format": "double"
					},
					"requirements": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"responsibilities": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"skills": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"description",
					"type",
					"field"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_InternshipBody_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"description": "ACADEMIC | PROFESSIONAL"
					},
					"field": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"isRemote": {
						"type": "boolean"
					},
					"isPaid": {
						"type": "boolean"
					},
					"stipendAmount": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string"
					},
					"durationWeeks": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"applicationDeadline": {
						"type": "string"
					},
					"slots": {
						"type": "number",
						"format": "double"
					},
					"requirements": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"responsibilities": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"skills": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Internship Monitoring & Connecting System API",
		"version": "1.0.0",
		"description": "Full-stack platform bridging students, companies, and supervisors throughout the internship lifecycle.",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "API Support"
		}
	},
	"paths": {
...
		"/api/logbook/me": {
			"get": {
				"operationId": "GetMyEntries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get all logbook entries for the authenticated student",
				"tags": [
					"Logbook"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/logbook": {
			"post": {
				"operationId": "CreateEntry",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Create a new logbook entry (Student only).\r\nFile attachment is handled via multipart/form-data in the actual endpoint.",
				"tags": [
					"Logbook"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LogbookEntryBody"
							}
						}
					}
				}
			}
		},
		"/api/logbook/{id}": {
			"put": {
				"operationId": "UpdateEntry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Update an existing logbook entry (Student only)",
				"tags": [
					"Logbook"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_LogbookEntryBody_"
							}
						}
					}
				}
			}
		},
		"/api/logbook/student/{studentId}": {
			"get": {
				"operationId": "GetStudentEntries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get logbook entries for a specific student (Supervisor / Admin)",
				"tags": [
					"Logbook"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "studentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/logbook/{id}/approve": {
			"patch": {
				"operationId": "ApproveEntry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Approve a logbook entry (Supervisor only)",
				"tags": [
					"Logbook"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ApproveEntryBody"
							}
						}
					}
				}
			}
		},
		"/api/logbook/attendance/me": {
			"get": {
				"operationId": "GetAttendanceSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get attendance summary for the authenticated student",
				"tags": [
					"Logbook"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/logbook/attendance": {
			"post": {
				"operationId": "LogAttendance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Log attendance for a student (Site Supervisor / Admin)",
				"tags": [
					"Logbook"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AttendanceBody"
							}
						}
					}
				}
			}
		},
		"/api/internships": {
			"get": {
				"operationId": "ListInternships",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Browse all open internship listings (public)",
				"tags": [
					"Internships"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "field",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "city",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isRemote",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "isPaid",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateInternship",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Create a new internship listing (Company only)",
				"tags": [
					"Internships"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InternshipBody"
							}
						}
					}
				}
			}
		},
		"/api/internships/{id}": {
			"get": {
				"operationId": "GetInternship",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get a single internship by ID (public)",
				"tags": [
					"Internships"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateInternship",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Update an internship listing (Company only)",
				"tags": [
					"Internships"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_InternshipBody_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteInternship",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Delete an internship listing (Company or Admin)",
				"tags": [
					"Internships"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/internships/{id}/bookmark": {
			"post": {
				"operationId": "ToggleBookmark",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Toggle bookmark on an internship (Student only)",
				"tags": [
					"Internships"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/internships/bookmarks/me": {
			"get": {
				"operationId": "GetBookmarks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get current student's bookmarked internships",
				"tags": [
					"Internships"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/internships/company/mine": {
			"get": {
				"operationId": "GetMyInternships",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get internships posted by the authenticated company",
				"tags": [
					"Internships"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/applications/internship/{internshipId}": {
			"get": {
				"operationId": "GetApplicationsForInternship",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get all applications for a specific internship (Company only)",
				"tags": [
					"Applications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "internshipId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}